Add router gateway support
When a gateway is set on router, the plugin dedicate a virtual network
as private network (called 'vn_snat_[ROUTER_UUID]') and start a netns
service instance between the private network and the public. Then it
sets a network policy between private and public network to pass through
the service instance. And finally, it creates a routing table with a
default route that point to the service instance private port.

Signed-off-by: Édouard Thuleau <edouard.thuleau@cloudwatt.com>